#32 - Began predicting attack strength using logged risk/reward vs observed risk/reward and available lethality.

+++ OpenRA.Mods.Common/AI/Esu/Database/SQLiteUtils.cs
@@ -49,11 +49,11 @@ namespace OpenRA.Mods.Common.AI.Esu.Database
             return sql;
         }
 
-        public static int GetCountForTable(SQLiteConnection openConnection, string tableName)
+        public static long GetCountForTable(SQLiteConnection openConnection, string tableName)
         {
             string sql = "SELECT COUNT(*) FROM " + tableName;
             SQLiteCommand countCommand = new SQLiteCommand(sql, openConnection);
-            return (int) countCommand.ExecuteScalar();
+            return (long) countCommand.ExecuteScalar();
         }
     }
